Working on reservation

+++ ikarimanov_ultraplay-cinemapi-2ac36e5d4762/CinemAPI.Models.Contracts/Projection/IProjection.cs
@@ -5,7 +5,7 @@ namespace CinemAPI.Models.Contracts.Projection
 {
     public interface IProjection
     {
-        long Id { get; }
+        int Id { get; }
 
         int RoomId { get; }
